--- input/form/login_user.go
@@ -1,6 +1,0 @@
-package form
-
-type LoginUserForm struct {
-	Email    string `form:"email" binding:"required"`
-	Password string `form:"password" binding:"required"`
-}
